requirements.txt
astroid==2.2.5
captcha2upload==0.2
certifi==2019.3.9
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
isort==4.3.17
jdcal==1.4.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
openpyxl==2.6.2
Pillow==6.0.0
pylint==2.3.1
pytesseract==0.2.6
requests==2.21.0
selenium==3.141.0
six==1.12.0
typed-ast==1.3.5
urllib3==1.24.2
wrapt==1.11.1
